% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avengers_movies.R
\name{avengers_movies}
\alias{avengers_movies}
\title{This function provides the list of all the Avengers movies to the user.}
\usage{
avengers_movies()
}
\value{
A dataframe that contains information of all the Avengers movies.
Dataframe contains Title, Type, Year, Genre, Runtime (in minutes), Actors, Director, Awards, and IMDB Rating.
}
\description{
This function provides the list of all the Avengers movies to the user.
}
\examples{
avengers_movies()
}
