% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actors.R
\name{get_actors}
\alias{get_actors}
\title{This function returns the actors of Movies and TV shows based on search word(s) by user.}
\usage{
get_actors(search_text, type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Actors.
}
\description{
This function returns the actors of Movies and TV shows based on search word(s) by user.
}
\examples{
get_actors('Guardians')
get_actors('Guardians Galaxy', type = 'movie')
}
