% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_awards.R
\name{get_awards}
\alias{get_awards}
\title{This function returns the awards received by Movies and TV shows based on search word(s) by user.}
\usage{
get_awards(search_text = "", type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Awards.
}
\description{
This function returns the awards received by Movies and TV shows based on search word(s) by user.
}
\examples{
get_awards('Guardians')
get_awards('Guardians Galaxy', type = 'movie')
}
