% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_director.R
\name{get_director}
\alias{get_director}
\title{This function returns the director of Movies and TV shows based on search word(s) by user.}
\usage{
get_director(search_text, type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Director.
}
\description{
This function returns the director of Movies and TV shows based on search word(s) by user.
}
\examples{
get_director('Guardians')
get_director('Guardians Galaxy', type = 'movie')
}
