% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genre.R
\name{get_genre}
\alias{get_genre}
\title{This function returns the genre of Movies and TV shows based on search word(s) by user.}
\usage{
get_genre(search_text, type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Genre.
}
\description{
This function returns the genre of Movies and TV shows based on search word(s) by user.
}
\examples{
get_genre('Guardians')
get_genre('Guardians Galaxy', type = 'movie')
}
