% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poster.R
\name{get_poster}
\alias{get_poster}
\title{This function returns a link to the poster of Movies and TV shows based on search word(s) by user.}
\usage{
get_poster(search_text = "", type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Poster (link).
}
\description{
This function returns a link to the poster of Movies and TV shows based on search word(s) by user.
}
\examples{
get_poster('Guardians')
get_poster('Guardians Galaxy', type = 'movie')
}
