% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rating.R
\name{get_rating}
\alias{get_rating}
\title{This function returns the IMDB rating of Movies and TV shows based on search word(s) by user.}
\usage{
get_rating(search_text = "", type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and IMDB Rating.
}
\description{
This function returns the IMDB rating of Movies and TV shows based on search word(s) by user.
}
\examples{
get_rating('Guardians')
get_rating('Guardians Galaxy', type = 'movie')
}
