% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_runtime.R
\name{get_runtime}
\alias{get_runtime}
\title{This function returns the runtime of Movies and TV shows based on search word(s) by user.}
\usage{
get_runtime(search_text, type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Runtime (in minutes).
}
\description{
This function returns the runtime of Movies and TV shows based on search word(s) by user.
}
\examples{
get_runtime('Guardians')
get_runtime('Guardians Galaxy', type = 'movie')
}
