% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top10_recommendations.R
\name{get_top10_recommendations}
\alias{get_top10_recommendations}
\title{This function returns the attributes of Top 10 most popular movies and TV shows by Genre.}
\usage{
get_top10_recommendations(genre = "adventure")
}
\arguments{
\item{genre}{Genre of recommendations can be - action, adventure, comedy, romance, drama, scifi, horror, animation
Default genre is adventure when there is no input from user.}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, Genre, Runtime (in minutes), Actors, Director, Awards, and IMDB Rating.
}
\description{
This function returns the attributes of Top 10 most popular movies and TV shows by Genre.
}
\examples{
get_top10_recommendations()
get_top10_recommendations(genre = 'drama')
}
