% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_all.R
\name{search_all}
\alias{search_all}
\title{This function returns the attributes of Movies and TV shows based on search word(s) by user.}
\usage{
search_all(search_text = "", type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, Genre, Runtime (in minutes), Actors, Director, Awards, and IMDB Rating.
}
\description{
This function returns the attributes of Movies and TV shows based on search word(s) by user.
}
\examples{
search_all('Guardians')
search_all('Guardians Galaxy', type = 'movie')
}
