% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprise_me.R
\name{surprise_me}
\alias{surprise_me}
\title{This function provides surprise recommendations of movies and TV shows to the user.}
\usage{
surprise_me(genre = "all")
}
\arguments{
\item{genre}{Genre of recommendations can be - action, adventure, comedy, romance, drama, scifi, horror, animation
When there is no input genre from user, a random recommendation from each of the genres is given.}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, Genre, Runtime (in minutes), Actors, Director, Awards, and IMDB Rating.
}
\description{
This function provides surprise recommendations of movies and TV shows to the user.
}
\examples{
surprise_me()
surprise_me(genre = 'adventure')
}
